Various updates.
Playing with captions for images.  it's not working well.

Moved the section on normalized device coordinates.

Changed demo number to chapter number to minimize confusion.

Added unwise rotate to demonstrate that a different method of thinking is required.

Negated the signs of the rotations, as it makes more sense.

Fixed tables to Teresa's liking.