Added in case for covariance.
This illustrates that thinking of a moving coordinate
system by reading the operations backwards doesn't always
work.

I've found that if I actively want to move the origin
somewhere else, or move the local basis somewhere else,
then reading the transformations backwards, as moving
set of axises is beneficial.

If instead, like for covariance, if I'm moving the non-natural
basis to the natural basis, then I need to read the transformations
in the order in which they are applied.